PaymentController의 엔드포인트가 Swagger UI 문서에 표시될 example value 설정 & Payment와 Pledge의 amount 데이터 타입을 Long에서 BigDecimal로 변경

+++ src/main/java/io/github/crowdfund/domain/payment/Payment.java
@@ -2,6 +2,8 @@ package io.github.crowdfund.domain.payment;
 
 import org.springframework.data.annotation.Id;
 import org.springframework.data.relational.core.mapping.Table;
+
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 /**
@@ -20,7 +22,7 @@ public record Payment(
     @Id Long id,
     Long pledgeId,
     String paymentMethod,
-    Long amount,
+    BigDecimal amount,
     String status,
     LocalDateTime paidAt,
     LocalDateTime createdAt
