+++ src/main/java/io/github/crowdfund/feature/pledges/user/UserPledgeService.java
@@ -55,7 +55,7 @@ public class UserPledgeService {
                 userId,
                 request.project_id(),
                 request.reward_id(),
-                reward.price().longValue(),
+                reward.price(),
                 FulfillmentStatus.READY,
                 null,
                 LocalDateTime.now()
